validator to support create and update
Fix interesting issue that server was not handling messages in go routine.
This meant the server calling the controller update validator, which in turn calls
the store, resulted in a timeout.
Store handlers now run in a go routine with a sync.WaitGroup to gracefull wait for
handlers to finish.

resolves #8