BREAKING: rename Account to Namespace
Based on experiences working with AWS, the word "account" is something that
continuously raises confusion.
Hence, KISS and copy kube lingo and call it a namespace.
We still have NATS accounts under the hood, but we do not need to expose that
complexity to users.

Also remove User and ServiceAccount objects because the design is wrong and
costing maintenance effort.